requirements.txt
streamlit
ultralytics
Pillow
groq
python-dotenv
